# FromNapiValue trait for parameters