--- a/lib/types/cql-utils.js
+++ b/lib/types/cql-utils.js
@@ -43,7 +43,7 @@ function guessTypeChecked(value) {
  * Wraps each of the elements into given subtype
  * @param {Array} values
  * @param {rust.ComplexType} subtype
- * @returns {Array<rust.QueryParameterWrapper>}
+ * @returns {Array<rust.ComplexType|any[]>} Returns tuple: [rust.ComplexType, any[]]
  */
 function encodeListLike(values, subtype) {
     if (!Array.isArray(values))
@@ -60,19 +60,19 @@ function encodeListLike(values, subtype) {
             values[i],
             "A collection can't contain null or unset values",
         );
-        res.push(wrapValueBasedOnType(subtype, values[i]));
+        res.push(wrapValueBasedOnType(subtype, values[i])[1]);
     }
-    return res;
+    return [subtype, res];
 }
 
 /**
  * @param {*} value
  * @param {rust.ComplexType} parentType
- * @returns {Array<Array<rust.QueryParameterWrapper>>}
+ * @returns {Array<rust.ComplexType|any[][]>} Returns tuple: [rust.ComplexType, any[][]]
  */
 function encodeMap(value, parentType) {
-    const keySubtype = parentType.getFirstSupportType();
-    const valueSubtype = parentType.getSecondSupportType();
+    let keySubtype = parentType.getFirstSupportType();
+    let valueSubtype = parentType.getSecondSupportType();
 
     let res = [];
 
@@ -83,12 +83,25 @@ function encodeMap(value, parentType) {
         const val = value[key];
         ensureValue(val, "A collection can't contain null or unset values");
         ensureValue(key, "A collection can't contain null or unset values");
+        if (!keySubtype || !valueSubtype) {
+            if (valueSubtype || keySubtype) {
+                throw new Error(
+                    `Internal error: Invalid support types for map`,
+                );
+            }
+            keySubtype = guessTypeChecked(key);
+            valueSubtype = guessTypeChecked(val);
+            parentType = parentType.remapMapSupportType(
+                keySubtype,
+                valueSubtype,
+            );
+        }
         res.push([
-            wrapValueBasedOnType(keySubtype || guessTypeChecked(key), key),
-            wrapValueBasedOnType(valueSubtype || guessTypeChecked(val), val),
+            wrapValueBasedOnType(keySubtype, key)[1],
+            wrapValueBasedOnType(valueSubtype, val)[1],
         ]);
     }
-    return res;
+    return [parentType, res];
 }
 
 /**
@@ -102,86 +115,95 @@ function ensureNumber(value) {
 }
 
 /**
- * Encodes tuple into QueryParameterWrapper
+ * Wraps tuple into format recognized by ParameterWrapper
  * @param {types.Tuple} value
  * @param {rust.ComplexType} type
- * @returns {Array<rust.QueryParameterWrapper>}
+ * @returns {Array<rust.ComplexType|Array<any>>} Returns tuple: [rust.ComplexType, Array<any>]
  */
 function encodeTuple(value, type) {
     let res = [];
+    // TODO:
+    // Add proper type guessing for tuples
+    // If some of the types are not provided (as it's a possible case in type guessing)
+    // then the returned type will have this information, while type provided in argument - not.
+    let newTypes = [];
     let types = type.getInnerTypes();
     for (let i = 0; i < types.length; i++) {
-        const element = value.get(i) !== undefined ? value.get(i) : null;
-        res.push(getWrapped(types[i], element));
+        const element = getWrapped(
+            types[i],
+            value.get(i) !== undefined ? value.get(i) : null,
+        );
+        newTypes.push(element[0]);
+        res.push(element[1]);
     }
-    return res;
+    return [rust.ComplexType.remapTupleSupportType(newTypes), res];
 }
 
 /**
- * Wrap value into MaybeUnsetQueryParameterWrapper based on the type
+ * Wraps value into format recognized by ParameterWrapper, based on the provided type
  * @param {rust.ComplexType} type
  * @param {*} value
- * @returns {rust.MaybeUnsetQueryParameterWrapper?}
+ * @returns {Array<rust.ComplexType|any>} Returns tuple: [rust.ComplexType, any] or []
  */
 function getWrapped(type, value) {
-    // Unsets were introduced in CQLv3, and the backend of the driver - Rust driver -
+    // Unset was introduced in CQLv3, and the backend of the driver - Rust driver -
     // works only with version >= 4 of CQL, so unset will always be supported.
     if (value === null) {
-        return null;
+        return [];
     } else if (value === types.unset) {
-        return rust.MaybeUnsetQueryParameterWrapper.unset();
+        return [undefined];
     }
-    return rust.MaybeUnsetQueryParameterWrapper.fromNonNullNonUnsetValue(
-        wrapValueBasedOnType(type, value),
-    );
+    return wrapValueBasedOnType(type, value);
 }
 
 /**
- * Wrap value, which is not Unset, into QueryParameterWrapper based on the type
+ * Wrap value, which is not Unset, into type and value pair,
+ * ensuring value is converted into expected form
  * @param {rust.ComplexType} type
  * @param {*} value
- * @returns {rust.QueryParameterWrapper}
+ * @returns {Array<rust.ComplexType|any>} Returns tuple: [rust.ComplexType, any]
  */
 function wrapValueBasedOnType(type, value) {
-    let encodedElement;
     let tmpElement;
     // To increase clarity of the error messages, in case value is of different type then expected,
     // when we call some methods on value variable, type is checked explicitly.
     // In other cases default Error will be thrown, which has message meaningful for the user.
     switch (type.baseType) {
+        // For these types, no action is required
         case rust.CqlType.Ascii:
-            return rust.QueryParameterWrapper.fromAscii(value);
-        case rust.CqlType.BigInt:
-            return rust.QueryParameterWrapper.fromBigint(
-                arbitraryValueToBigInt(value),
-            );
-        case rust.CqlType.Blob:
-            return rust.QueryParameterWrapper.fromBlob(value);
         case rust.CqlType.Boolean:
-            return rust.QueryParameterWrapper.fromBoolean(value);
+        case rust.CqlType.Blob:
+        case rust.CqlType.Decimal:
+        case rust.CqlType.Double:
+        case rust.CqlType.Empty:
+        case rust.CqlType.Float:
+        case rust.CqlType.Text:
+            break;
+        case rust.CqlType.BigInt:
+            value = arbitraryValueToBigInt(value);
+            break;
         case rust.CqlType.Counter:
-            return rust.QueryParameterWrapper.fromCounter(BigInt(value));
+            value = BigInt(value);
+            break;
         case rust.CqlType.Date:
             if (!(value instanceof LocalDate))
                 throw new TypeError(
                     "Expected LocalDate type to parse into Cql Date",
                 );
-            return rust.QueryParameterWrapper.fromLocalDate(
-                value.getInternal(),
-            );
-        case rust.CqlType.Double:
-            return rust.QueryParameterWrapper.fromDouble(value);
+            value = value.getInternal();
+            break;
         case rust.CqlType.Duration:
             if (!(value instanceof Duration))
                 throw new TypeError(
                     "Expected Duration type to parse into Cql Duration",
                 );
-            return rust.QueryParameterWrapper.fromDuration(value.getInternal());
-        case rust.CqlType.Float:
-            return rust.QueryParameterWrapper.fromFloat(value);
+            value = value.getInternal();
+            break;
         case rust.CqlType.Int:
+        case rust.CqlType.SmallInt:
+        case rust.CqlType.TinyInt:
             ensureNumber(value);
-            return rust.QueryParameterWrapper.fromInt(value);
+            break;
         case rust.CqlType.Set:
             // TODO:
             // This is part of the datastax logic for encoding set.
@@ -196,10 +218,11 @@ function wrapValueBasedOnType(type, value) {
                 });
                 return this.encodeList(arr, subtype);
             } */
-            encodedElement = encodeListLike(value, type.getFirstSupportType());
-            return rust.QueryParameterWrapper.fromSet(encodedElement);
-        case rust.CqlType.Text:
-            return rust.QueryParameterWrapper.fromText(value);
+            value = encodeListLike(value, type.getFirstSupportType());
+            if (!type.getFirstSupportType())
+                type = type.remapListSupportType(value[0]);
+            value = value[1];
+            break;
         case rust.CqlType.Timestamp:
             tmpElement = value;
             if (typeof value === "string") {
@@ -212,8 +235,8 @@ function wrapValueBasedOnType(type, value) {
                     throw new TypeError("Invalid date: " + tmpElement);
                 }
             }
-
-            return rust.QueryParameterWrapper.fromTimestamp(BigInt(value));
+            value = BigInt(value);
+            break;
         case rust.CqlType.Inet:
             // Other forms of providing InetAddress are kept as parity with old driver
             if (typeof value === "string") {
@@ -227,10 +250,14 @@ function wrapValueBasedOnType(type, value) {
                     "Expected InetAddress type to parse into Cql Inet",
                 );
             }
-            return rust.QueryParameterWrapper.fromInet(value.getInternal());
+            value = value.getInternal();
+            break;
         case rust.CqlType.List:
-            encodedElement = encodeListLike(value, type.getFirstSupportType());
-            return rust.QueryParameterWrapper.fromList(encodedElement);
+            value = encodeListLike(value, type.getFirstSupportType());
+            if (!type.getFirstSupportType())
+                type = type.remapListSupportType(value[0]);
+            value = value[1];
+            break;
         case rust.CqlType.Map:
             // TODO:
             // This is part of the datastax logic for encoding map.
@@ -242,11 +269,10 @@ function wrapValueBasedOnType(type, value) {
                 // Use Map#forEach() method to iterate
                 value.forEach(addItem);
             } */
-            encodedElement = encodeMap(value, type);
-            return rust.QueryParameterWrapper.fromMap(encodedElement);
-        case rust.CqlType.SmallInt:
-            ensureNumber(value);
-            return rust.QueryParameterWrapper.fromSmallInt(value);
+            value = encodeMap(value, type);
+            type = value[0];
+            value = value[1];
+            break;
         case rust.CqlType.Time:
             // Other forms of providing LocalTime are kept as parity with old driver
             if (typeof value == "string") {
@@ -258,12 +284,8 @@ function wrapValueBasedOnType(type, value) {
                 );
             }
 
-            return rust.QueryParameterWrapper.fromLocalTime(
-                value.getInternal(),
-            );
-        case rust.CqlType.TinyInt:
-            ensureNumber(value);
-            return rust.QueryParameterWrapper.fromTinyInt(value);
+            value = value.getInternal();
+            break;
         case rust.CqlType.Uuid:
             // Other forms of providing UUID are kept as parity with old driver
             if (typeof value === "string") {
@@ -275,10 +297,13 @@ function wrapValueBasedOnType(type, value) {
                     "Expected UUID type to parse into Cql Uuid",
                 );
             }
-            return rust.QueryParameterWrapper.fromUuid(value.getInternal());
+            value = value.getInternal();
+            break;
         case rust.CqlType.Tuple:
-            encodedElement = encodeTuple(value, type);
-            return rust.QueryParameterWrapper.fromTuple(encodedElement);
+            value = encodeTuple(value, type);
+            type = value[0];
+            value = value[1];
+            break;
         case rust.CqlType.Timeuuid:
             // Other forms of providing TimeUUID are kept as parity with old driver
             if (typeof value === "string") {
@@ -289,26 +314,28 @@ function wrapValueBasedOnType(type, value) {
                     "Expected Time UUID type to parse into Cql Uuid",
                 );
             }
-            return rust.QueryParameterWrapper.fromTimeUuid(value.getInternal());
+            value = value.getInternal();
+            break;
         default:
             // Or not yet implemented type
             throw new ReferenceError(
                 `[INTERNAL DRIVER ERROR] Unknown type: ${type}`,
             );
     }
+    return [type, value];
 }
 
 /**
- * Parses array of params into rust objects according to preparedStatement expected types
- * Throws ResponseError when received different amount of parameters than expected
+ * Parses array of params into expected format according to preparedStatement expected types
  *
  * If `allowGuessing` is true, then for each missing field of `expectedTypes`, this function will try to guess a type.
  * If the type cannot be guessed, error will be thrown.
- * Field is missing if it is null, undefined (or if the `expectedTypes` list is to short)
+ * Field is missing if it is null, undefined or if the `expectedTypes` list is too short
  * @param {Array<rust.ComplexType | null>} expectedTypes List of expected types.
  * @param {Array<any>} params
  * @param {boolean} [allowGuessing]
- * @returns {Array<rust.QueryParameterWrapper>}
+ * @returns {Array<rust.ComplexType|any>} Returns tuple: [rust.ComplexType, any] or []
+ * @throws ResponseError when received different amount of parameters than expected
  */
 function parseParams(expectedTypes, params, allowGuessing) {
     if (expectedTypes.length == 0 && !params) return [];
@@ -327,8 +354,10 @@ function parseParams(expectedTypes, params, allowGuessing) {
             params[i] = null;
         }
 
-        if (params[i] === null) {
-            res.push(null);
+        // undefined as null depends on encodingOptions.useUndefinedAsUnset option
+        // TODO: Add support for this option
+        if (params[i] === null || params[i] === undefined) {
+            res.push([]);
             continue;
         }