--- a/src/tests/request_values_tests.rs
+++ b/src/tests/request_values_tests.rs
@@ -8,7 +8,7 @@ use std::{
 };
 
 use crate::{
-    requests::parameter_wrappers::QueryParameterWrapper,
+    requests::parameter_wrappers::ParameterWrapper,
     types::type_wrappers::{ComplexType, CqlType},
 };
 
@@ -38,7 +38,7 @@ pub fn tests_from_value_get_type(test: String) -> ComplexType {
         "Time" => (CqlType::Time, None, None),
         "Timeuuid" => (CqlType::Timeuuid, None, None),
         "Tuple" => {
-            return ComplexType::from_tuple(&[
+            return ComplexType::tuple_from_column_type(&[
                 ColumnType::Native(NativeType::Text),
                 ColumnType::Tuple(vec![
                     ColumnType::Native(NativeType::Int),
@@ -58,7 +58,7 @@ pub fn tests_from_value_get_type(test: String) -> ComplexType {
 }
 
 #[napi]
-pub fn tests_from_value(test: String, value: &QueryParameterWrapper) {
+pub fn tests_from_value(test: String, value: ParameterWrapper) {
     let v = match test.as_str() {
         "Ascii" => CqlValue::Ascii("Some arbitrary value".to_owned()),
         "BigInt" => CqlValue::BigInt(i64::MAX),
@@ -103,5 +103,16 @@ pub fn tests_from_value(test: String, value: &QueryParameterWrapper) {
         "Uuid" => CqlValue::Uuid(uuid!("ffffffff-eeee-ffff-ffff-ffffffffffff")),
         _ => CqlValue::Empty,
     };
-    assert_eq!(value.parameter, v);
+    assert_eq!(
+        match value.row {
+            Some(v) => {
+                match v {
+                    scylla::value::MaybeUnset::Unset => panic!("Expected set value"),
+                    scylla::value::MaybeUnset::Set(w) => w,
+                }
+            }
+            None => panic!("Expected some value"),
+        },
+        v
+    );
 }