Before, we created QueryParameterWrapper with from_... functions, and then cloning this object when executing query.

Now, with implementation of FromNapiValue trait, we pass parameters to queries directly, without any additional conversion steps. In order to correctly handle value types, each parameter is passed as a pair (type, value). Information about type is then used on the rust side to create correct CqlValue object.

This significantly improves performance for concurrent inserts benchmark, and slightly for plain inserts benchmark.

Concurrent inset. Around 30% improvement in time elapsed, 20% in task clock

After:

```bash
sudo perf stat node ./benchmark/logic/concurrent_insert.js scylladb-javascript-driver 1000000

 Performance counter stats for 'node ./benchmark/logic/concurrent_insert.js scylladb-javascript-driver 1000000':

         70,343.46 msec task-clock                       #    1.892 CPUs utilized             
         2,909,393      context-switches                 #   41.360 K/sec                     
            40,946      cpu-migrations                   #  582.087 /sec                      
           190,074      page-faults                      #    2.702 K/sec                     
   273,230,048,329      cycles                           #    3.884 GHz                       
    72,133,512,748      stalled-cycles-frontend          #   26.40% frontend cycles idle      
   205,423,999,244      instructions                     #    0.75  insn per cycle            
                                                  #    0.35  stalled cycles per insn   
    39,462,664,735      branches                         #  561.000 M/sec                     
     1,481,648,921      branch-misses                    #    3.75% of all branches           

      37.183329662 seconds time elapsed

      46.102799000 seconds user
      25.690728000 seconds sys
```

Before

```bash
sudo perf stat node ./benchmark/logic/concurrent_insert.js scylladb-javascript-driver 1000000

 Performance counter stats for 'node ./benchmark/logic/concurrent_insert.js scylladb-javascript-driver 1000000':

         86,873.69 msec task-clock                       #    1.633 CPUs utilized             
         3,221,740      context-switches                 #   37.085 K/sec                     
            48,446      cpu-migrations                   #  557.660 /sec                      
           190,596      page-faults                      #    2.194 K/sec                     
   344,373,066,097      cycles                           #    3.964 GHz                       
    87,340,988,626      stalled-cycles-frontend          #   25.36% frontend cycles idle      
   246,727,230,093      instructions                     #    0.72  insn per cycle            
                                                  #    0.35  stalled cycles per insn   
    47,143,750,188      branches                         #  542.670 M/sec                     
     1,750,876,479      branch-misses                    #    3.71% of all branches           

      53.200649700 seconds time elapsed

      60.397169000 seconds user
      28.058204000 seconds sys


```

Plain insert. Around 10% improvement in user time and number of instructions, 90% drop in page faults!

After:

```bash
sudo perf stat node ./benchmark/logic/insert.js scylladb-javascript-driver 300000

 Performance counter stats for 'node ./benchmark/logic/insert.js scylladb-javascript-driver 300000':

         30,435.03 msec task-clock                       #    0.839 CPUs utilized             
         2,126,428      context-switches                 #   69.868 K/sec                     
            19,269      cpu-migrations                   #  633.119 /sec                      
             8,034      page-faults                      #  263.972 /sec                      
   104,314,726,271      cycles                           #    3.427 GHz                       
    51,460,535,207      stalled-cycles-frontend          #   49.33% frontend cycles idle      
    91,254,504,996      instructions                     #    0.87  insn per cycle            
                                                  #    0.56  stalled cycles per insn   
    17,802,023,048      branches                         #  584.919 M/sec                     
     1,010,940,030      branch-misses                    #    5.68% of all branches           

      36.283305820 seconds time elapsed

      14.569449000 seconds user
      17.101418000 seconds sys
```

Before:

```bash
sudo perf stat node ./benchmark/logic/insert.js scylladb-javascript-driver 300000

 Performance counter stats for 'node ./benchmark/logic/insert.js scylladb-javascript-driver 300000':

         31,847.19 msec task-clock                       #    0.830 CPUs utilized             
         2,126,649      context-switches                 #   66.777 K/sec                     
            18,765      cpu-migrations                   #  589.220 /sec                      
            88,511      page-faults                      #    2.779 K/sec                     
   111,466,616,472      cycles                           #    3.500 GHz                       
    51,182,304,492      stalled-cycles-frontend          #   45.92% frontend cycles idle      
   100,318,097,101      instructions                     #    0.90  insn per cycle            
                                                  #    0.51  stalled cycles per insn   
    19,409,416,734      branches                         #  609.455 M/sec                     
     1,016,535,128      branch-misses                    #    5.24% of all branches           

      38.359245639 seconds time elapsed

      16.112863000 seconds user
      16.940505000 seconds sys
```